requirements.txt
absl-py==0.12.0
autopep8==1.6.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.11.0
fonttools==4.28.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.5.0
numpy==1.20.3
oauthlib==3.1.0
packaging==21.3
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-scm==6.3.2
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
toml==0.10.2
tomli==1.2.2
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
